1.0.0 recommit
Version 1.0.0 (5 may 2015)
The password is stored in keyring now (thanks to JBu92 for the
suggestion)
If there's a password in settings.xml, it will be converted to the
password in keyring (the password in settings.xml will be blanked).
If the user wants to view a sponsored video and the username/password
doesn't work, the user will be asked to type in the correct password. If
the userid/password works, the password will be stored in keyring.
Added a button in settings to force the removal of the password from
keyring
Switched from LWPCookieJar to MozillaCookieJar due to bug which caused
sponsored vids not working
Always delete cookie file
Small changes in logging
Sponsor Early Screening videos are NOT working