1.3.11-SNAPSHOT (2019-05-12)
- added adapted inputstream switch. When this is on, the videoplayer inputstream addon 'Inputstream Adaptive'
(kudo's to peak3d) is used. The addon 'Inputstream Adaptive' determines what resolution it will display based upon the
available bandwidth and the settings of the addon 'Inputstream Adaptive'.
When the adapted inputstream switch is off, the roosterteeth addon will try and play the video based upon the
maximum video resolution in the roosterteeth settings.
- for when the adapted inputstream switch is off: if a video of the desired video is not found, look for a video of lower quality